Chromium flavor implementations
