+++ src/platforms/web/chromeos/root/util.js
@@ -1,1 +1,5 @@
-../../util.js
+// tiny util.inherits shim.
+exports.inherits = function (Class, Parent) {
+  Class.prototype.__proto__ = Parent.prototype;
+};
+
