+++ src/platforms/web/chromium/crx/background.html
@@ -4,6 +4,6 @@
         <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
         <title>Snap4Arduino Plugin</title>
         <link rel="shortcut icon" href="s4a.ico" type="image/gif">
-        <script autorequire="/main.js" moduledir="/node_modules/" src="/node_modules/chrome-app-module-loader/module.js" ></script>
+        <script src="bundle.js" ></script>
     </head>
 </html>
