--- src/platforms/web/chromium/crx/events.js
@@ -1,1 +1,0 @@
-../../events.js
