--- src/platforms/web/chromium/crx/firmata.js
@@ -1,1 +1,0 @@
-../../firmata.js
