+++ src/platforms/web/chromium/crx/manifest.json
@@ -24,5 +24,6 @@
         ],
         "accepts_tls_channel_id": false
     },
-    "icons": { "16": "icon-16.png", "128": "icon-128.png" }
+    "icons": { "16": "icon-16.png", "128": "icon-128.png" },
+    "key": "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAryxXD9KBQe4/H/Xqpbj7Uz8V4pXtst4jb9olXllFRWQJbOx4OkTPkiF9Fmm5q4HCldD/XVuKOCGbiafdMlrZI2iRT2Jwr9UAf1VB0ajiqTUrSxzfn3Aiwa8LaiiAI3krmD8QyoCZTuxTsyIUMXCzIf4pVmgzanWmHX+lfu1TExXrmzsFy0BLOWFFJ8VPs/ArhoqoNRw3oLXczXcjbyRDMyhuj6/GYCJXOy/wwINLsXnJdhsyPPbsX3cin/tWnqb3a2ElY/tUsJdqK0RRF7aqJOLYIzWbegIQRcUsUwkTgZRaJkBbH+dLdG7J2yici2JZ12882mdU4mRSYOnuGN+lIwIDAQAB"
 }
