--- src/platforms/web/chromium/crx/serialport.js
@@ -1,1 +1,0 @@
-../../serialport.js
