--- src/platforms/web/chromium/crx/util.js
@@ -1,1 +1,0 @@
-../../util.js
