--- src/platforms/web/util.js
@@ -1,5 +1,0 @@
-// tiny util.inherits shim.
-exports.inherits = function (Class, Parent) {
-  Class.prototype.__proto__ = Parent.prototype;
-};
-
