+++ src/s4a/arduino.js
@@ -330,6 +330,8 @@ Arduino.prototype.connect = function (port, verify, channel) {
                 }
 
                 myself.hideMessage();
+                myself.board.getArduinoBoardParam = nop;
+
                 ide.inform(myself.owner.name, localize('An Arduino board has been connected. Happy prototyping!'));   
             } else {
                 fail(err);
